% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{scMetaTraj_axis_score}
\alias{scMetaTraj_axis_score}
\title{Calculate metabolic axis scores}
\usage{
scMetaTraj_axis_score(scores, axis_list, scale = TRUE)
}
\arguments{
\item{scores}{Cell-by-module metabolic score matrix}

\item{axis_list}{Named list defining metabolic axes}

\item{scale}{Logical, whether to scale axis scores (Z-score)}
}
\value{
A matrix of cell-by-axis scores
}
\description{
Calculate metabolic axis scores
}
