% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gradient.R
\name{scMetaTraj_plot_gradient}
\alias{scMetaTraj_plot_gradient}
\title{Plot metabolic gradient on UMAP}
\usage{
scMetaTraj_plot_gradient(
  embedding,
  score,
  title = "Metabolic gradient",
  palette
)
}
\arguments{
\item{embedding}{UMAP coordinates (data.frame or matrix)}

\item{score}{Vector of metabolic axis score}

\item{title}{Plot title}

\item{palette}{Continuous color palette}
}
\value{
ggplot object
}
\description{
Plot metabolic gradient on UMAP
}
