% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_module_cards.R
\name{scMetaTraj_plot_module_cards}
\alias{scMetaTraj_plot_module_cards}
\title{Plot metabolic module cards}
\usage{
scMetaTraj_plot_module_cards(cluster_profile, module_axis_map, axis_palette)
}
\arguments{
\item{cluster_profile}{Cluster-by-module matrix}

\item{module_axis_map}{Data.frame with columns: module, axis}

\item{axis_palette}{Named vector of colors for axes}
}
\value{
ggplot object
}
\description{
Plot metabolic module cards
}
