% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend_by_cluster.R, R/trend_by_cluster.R
\name{scMetaTraj_plot_trend_by_cluster}
\alias{scMetaTraj_plot_trend_by_cluster}
\title{Plot module trends by metabolic cluster}
\usage{
scMetaTraj_plot_trend_by_cluster(trend_by_cluster, palette)

scMetaTraj_plot_trend_by_cluster(trend_by_cluster, palette)
}
\arguments{
\item{trend_by_cluster}{Output from scMetaTraj_trend_by_cluster().}

\item{palette}{Named vector of cluster colors (e.g., scMetaTraj_palette_discrete).}
}
\value{
ggplot object.

ggplot object.
}
\description{
Plot module trends by metabolic cluster

Plot module trends along mPT stratified by cluster
}
