% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_multi.R
\name{scMetaTraj_plot_trend_multi}
\alias{scMetaTraj_plot_trend_multi}
\title{Plot module trends along mPT with switchpoints}
\usage{
scMetaTraj_plot_trend_multi(trend_long, switchpoints)
}
\arguments{
\item{trend_long}{Output $trend_long from scMetaTraj_trend_multi().}

\item{switchpoints}{Output $switchpoints from scMetaTraj_trend_multi().}
}
\value{
ggplot object.
}
\description{
Plot module trends along mPT with switchpoints
}
